Did some serious reworking and cleaning of the file structures. Split up the main app class
