--- updater.py
@@ -1,5 +1,0 @@
-
-#This is a seperate program used to update the software
-#Gather saved data from previous version, if installed local or admin?
-#Remove old version
-#Install new version
